+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add a `main_class` getter to namespaces
 * Add `String#countCharacters()` to count characters, including emojis
 * Make `String#count()` use `String#countCharacters()` when empty string is given
+* Add `String#substrCharacters(begin, length)` to get a substring of the actual characters
 
 ## 0.6.4 (2019-01-12)